requirements.txt
blinker==1.7.0
click==8.1.7
dnspython==1.16.0
Flask==3.0.2
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==23.2
pymongo==3.11.1
python-dotenv==1.0.1
Werkzeug==3.0.1